requirements.txt
astroid==2.2.5
autopep8==1.4.4
Babel==2.7.0
beautifulsoup4==4.8.1
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cssselect2==0.2.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.2.6
django-autocomplete-light==3.5.1
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-crispy-forms==1.7.2
django-currencies==0.9.0
django-filter==2.3.0
django-form-utils==1.0.3
django-heroku==0.3.1
django-import-export==2.3.0
django-js-asset==1.2.2
django-model-utils==3.2.0
django-money==0.15.1
django-mptt==0.11.0
django-multiselectfield==0.1.12
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
install==1.3.3
isort==4.3.21
jdcal==1.4.1
jsonfield==2.0.2
lazy-object-proxy==1.4.2
MarkupPy==1.14
mccabe==0.6.1
money==1.3.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==6.2.0
pkg-resources==0.0.0
ply==3.11
progressbar2==3.47.0
psycopg2==2.8.5
py-moneyed==0.8.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-utils==2.3.0
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.32
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
swapper==1.1.1
tablib==2.0.0
thriftpy2==0.4.12
tinycss2==1.0.2
tqdm==4.36.1
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.6
wadofstuff-django-serializers==1.1.0
WeasyPrint==51
webencodings==0.5.1
whitenoise==3.2
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zk==0.6.10
